requirements.txt
reportlab
Pillow
requests
beautifulsoup4  # needed for enrich_metadata.py
opencv-python  # needed for video_to_images.py
google-generativeai  # neded for image_descriptions.py
